requirements.txt
faust-streaming==0.10.1
motor==3.1.1
diagrams